Some modification

+++ libft/ft_strcmp.c
@@ -6,7 +6,7 @@
 /*   By: kedavain <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/03/27 14:11:00 by kedavain          #+#    #+#             */
-/*   Updated: 2021/04/02 17:41:53 by kedavain         ###   ########.fr       */
+/*   Updated: 2021/04/04 17:24:36 by kedavain         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 int	ft_strcmp(char *s1, char *s2)
 {
-	int		i;
+	unsigned int		i;
 
 	i = 0;
 	while (s1[i] == s2[i] && s1[i] && s2[i])
 		i++;
-	return ((unsigned)s1[i] - (unsigned)s2[i]);
+	return (s1[i] - s2[i]);
 }
